requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.33
langchain_google_genai==1.0.9
numpy==2.0.1
opencv_python==4.10.0.84
Pillow==10.4.0
protobuf==3.20.3
PyPDF2==3.0.1
pytesseract==0.3.10
scikit_learn==1.4.2
streamlit==1.32.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
